User Profiles Loading, Correct Layout Admin/Non
Everything is loading in in a timely fashion, not the best way, but it
works for now. Using a buffer system where the group / user profiles
load in as clicked and are ready by the time the page loads. I also went
thru and made all the layouts display correctly based on the level of
AUTH the current user has for each page.