Generalizing List Layouts / Activities
Also made it so you can only invite members to a group if you are in the
group, will need to do a check later for the admin level of the user.
Still need to spend more time cleaning up listActivity, it is still very
crowded and can be reduced as well as adding requests / invites to it.
Still haven't put in loading popups / foregrounds to let the user know a
new page is loading. Will continue later, and all day tomorrow before
meeting.