Refactored Profiles Together, Entity Inheritance
Did a lot of refactoring to reduce the # of files. Getting pretty clean
now but will need to do some more work on it. User profiles and group
profiles use the same class, just use a intent extra "content" to
differentiate. I also generalized a lot of the layouts so that they can
be reused. Also made group and user inherit from entity, and will do the
same for event tomorrow.