Fixed Button Actions in ListActivity
Pretty stable version, still need to clean out the addnew button on 1
or 2 of the lists where it shouldn’t be and next make sure to check for
admin status before giving the user the ability to add new groups /
members. On that same token, will be moving into checking admin for
edit profiles.