startDate/endDate default values and toBringList
-startDate/endDate in createEventActivity and editEventActivity load
default values based on the previously set dates in the editText.
endDate also defaults to startDate if not set.

-bugs fixed with createEvent pre-checks

-toBringList implemented in createEventActivity.  sends array of items
to server and stores in db table.  Still needs some work on the
toBringList popup window.  (toBring ArrayList currently does not load
previously typed items if toBring button is hit a second time, etc.)

-PHPs modified slightly to receive new inputs.