Updated List Layout, Setup Message Class
Looking to smartly create a message activity that can handle real time
messaging with groups, events and users. Got a message class to hold
all stuff pertinent, will have array list inside of the messaging
activity and will be utilizing the broadcast receivers for realtime
updates. Looking to work with todd on that.