Full implementation of changePassword, changeEmail, and deleteAccount
changePassword - requires old password to be typed.  generates a new
hash in the database.
changeEmail - requires old password to be typed, requires new email that
is not in use.  Updates all client-side user values as well as database.
deleteAccount - flags account for deletion.  account is still available
for 30 days until mysqlscheduler deletes it.  If user logs back in
within 30 days, account flag is removed and account is active status
again.